AbstractSolutionGeneratedHandler, SolutionGeneratedEvent Commented VND pending reimplement Use seconds instead of milliseconds by default, use doubles instead of longs to avoid precission loss.

+++ src/main/java/es/urjc/etsii/grafo/io/WorkingOnResult.java
@@ -50,8 +50,8 @@ public class WorkingOnResult<S extends Solution<I>, I extends Instance> {
         }
     }
 
-    public long getAverageExecTimeInSeconds() {
-        long totalTime = 0;
+    public double getAverageExecTimeInSeconds() {
+        double totalTime = 0;
         for (var solution : this.solutions) {
             totalTime += solution.getExecutionTimeInNanos();
         }
